# Optimizations to preload nodes in history proofs