We were doing a LOT of direct get requests to the data layer to retrieve the various nodes needed to do proof generation. This can be demonstrated with the benchmark added in this PR, and additionally artifically slowing the database's `get` call to simulate some network latency.

```bash
$ cargo bench -p akd --bench directory -F bench,slow_internal_db
...
Benchmark key history proof generation on a small tree
                        time:   [4.2915 s 4.3005 s 4.3096 s]
```

Then if we enable the support for preloading nodes in `akd/src/directory.rs` that this PR adds, we see that we are just calling batch_get which greatly improves our overall performance, even if batch-get has similar latency

```bash
$ cargo bench -p akd --bench directory -F bench,slow_internal_db,preload_history
...
Benchmark key history proof generation on a small tree
                        time:   [1.7625 s 1.7663 s 1.7700 s]
                        change: [-59.051% -58.929% -58.804%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Without any artifical db slowing we have the following:

No preload:
```bash
Benchmark key history proof generation on a small tree
                        time:   [23.138 ms 23.557 ms 23.977 ms]
```

With preload:
```bash
Benchmark key history proof generation on a small tree
                        time:   [23.798 ms 24.204 ms 24.608 ms]
```