@slawlor 2023-01-17
Current metric with "slow db" is 1.72s / proof. updated benchmark for faster speedup in data generation