@eozturk1 2023-01-17
Could we get the database metrics to see whether we still have other single get requests so that we have a sense of there is more optimizations to be done here? (Feel free to merge this one. If there is room for improvement, we can create a separate task for it.)