@eozturk1 2023-01-17
> > Could we get the database metrics to see whether we still have other single get requests so that we have a sense of there is more optimizations to be done here? (Feel free to merge this one. If there is room for improvement, we can create a separate task for it.)
> 
> Yeah so it's a bit of a hacky test, but i threw together something by editing the test `test_simple_key_history` which is generating a series of key history proofs. We start with
> 
> 1. Editing the storage manager to have a cache
> 2. Flushing the cache after the writes (to simulate an empty cache)
> 3. Generating our proofs
> 4. Adding a `storage.log_metrics` at the end of the function then panicking lol
> 
> There's 2 GET requests which are constant from the Directory's azks retrieval and first root node retrieval ops, but the rest are batch-get's with 0 additional gets.
> 
> ```
> 00:00:00.034] INFO   Preload of tree (31 nodes) completed (append_only_zks:539)
> [00:00:00.034] INFO   Key history preloaded 31 nodes (directory:477)
> [00:00:00.077] INFO   Preload of tree (25 nodes) completed (append_only_zks:539)
> [00:00:00.077] INFO   Key history preloaded 25 nodes (directory:477)
> [00:00:00.101] INFO   Preload of tree (25 nodes) completed (append_only_zks:539)
> [00:00:00.101] INFO   Key history preloaded 25 nodes (directory:477)
> [00:00:00.120] INFO   Preload of tree (17 nodes) completed (append_only_zks:539)
> [00:00:00.120] INFO   Key history preloaded 17 nodes (directory:477)
> [00:00:00.136] ERROR  Cache hit since last: 604, cached size: 35 items (high_parallelism:62)
> [00:00:00.136] ERROR  Transaction writes: 74, Transaction reads: 63 (transaction:88)
> [00:00:00.136] ERROR  
> ===================================================
> ============ Database operation counts ============
> ===================================================
>     SET 2, 
>     BATCH SET 6, 
>     GET 2, 
>     BATCH GET 8
>     TOMBSTONE 0
>     GET USER STATE 0
>     GET USER DATA 4
>     GET USER STATE VERSIONS 6
> ===================================================
> ============ Database operation timing ============
> ===================================================
>     TIME READ 0 ms
>     TIME WRITE 0 ms (manager:175)
> ```
> 
> Without preloads we have:
> 
> ```
> [00:00:00.107] ERROR  Cache hit since last: 506, cached size: 35 items (high_parallelism:62)
> [00:00:00.107] ERROR  Transaction writes: 74, Transaction reads: 63 (transaction:88)
> [00:00:00.107] ERROR  
> ===================================================
> ============ Database operation counts ============
> ===================================================
>     SET 2, 
>     BATCH SET 6, 
>     GET 37, 
>     BATCH GET 0
>     TOMBSTONE 0
>     GET USER STATE 0
>     GET USER DATA 4
>     GET USER STATE VERSIONS 6
> ===================================================
> ============ Database operation timing ============
> ===================================================
>     TIME READ 0 ms
>     TIME WRITE 0 ms (manager:175)
> ```

Methodology SGTM! Hmm.. it is odd that we don't have any GETs for the future proofs -- as in until the next marker. See https://github.com/facebook/akd/blob/main/akd/src/directory.rs#L511. Maybe we the last version is equal to the next marker in the test code? IIRC, the latest version is the number of times the key has been changed, so the proofs until the next marker isn't maybe too many?